local-cluster: move alpenglow finality tests into local_cluster.rs; trim runtime
They lived in a separate alpenglow_finality.rs binary that the --test local_cluster CI lanes never picked up, so they weren't running in CI. Moved into local_cluster.rs so they hash-partition across the lanes instead of stacking into one serial block; tightened observation windows and the Alpenglow warmup (~33% less absolute time, ~65s/test). fmt + clippy clean; all 4 pass. Assertions, node counts, and stake ratios unchanged.
